+++ src/win32/amd64/aes-ni/aes.c
@@ -22,7 +22,7 @@ extern void aes_inverse_ASM(void *block, const void *key, uint64_t rounds);
 struct AES_STATE
 {
     unsigned char key[336];
-    size_t rounds;
+    unsigned rounds;
 };
 #endif /* annotation */
 
@@ -157,13 +157,13 @@ static const uint8_t ks[11] =
 };
 
 static void ExpandKey(const uint8_t *key, uint8_t *ext,
-                      size_t key_len, size_t rounds)
+                      size_t key_len, unsigned rounds)
 {
     size_t t;
 
     memcpy(ext, key, key_len * 4);
 
-    for (t = key_len; t < 4 * (rounds + 1); ++t)
+    for (t = key_len; t < (size_t)(4 * (rounds + 1)); ++t)
     {
         uint8_t tmp[5];
